backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
python-multipart
